fix: support decimal values and SDK user agents

+++ src/Models/Messenger/SendBulkMessengerResponse.php
@@ -15,7 +15,7 @@ final class SendBulkMessengerResponse
      * that specific receptor failed (`errorCode` set, `messageStatus` null).
      * `statusCode` always holds the raw value so unknown future codes never crash decoding.
      *
-     * @param array<int, array{message_id: ?string, receptor: string, local_id: ?string, hide: bool, statusCode: int, messageStatus: ?WebServiceMessageStatus, errorCode: ?WebServiceResponseCode, cost: int}> $receptors
+     * @param array<int, array{message_id: ?string, receptor: string, local_id: ?string, hide: bool, statusCode: int, messageStatus: ?WebServiceMessageStatus, errorCode: ?WebServiceResponseCode, cost: float}> $receptors
      * @param array<int, int> $counts
      */
     public function __construct(
@@ -24,7 +24,7 @@ final class SendBulkMessengerResponse
         public readonly string $message,
         public readonly ?\DateTimeImmutable $sendTime,
         public readonly int $totalCount,
-        public readonly int $totalCost,
+        public readonly float $totalCost,
         public readonly array $counts,
         public readonly string $profile,
         public readonly string $messenger,
@@ -56,7 +56,7 @@ final class SendBulkMessengerResponse
                         'statusCode' => $statusCode,
                         'messageStatus' => WebServiceMessageStatus::tryFrom($statusCode),
                         'errorCode' => WebServiceResponseCode::tryFrom($statusCode),
-                        'cost' => (int) $receptor['cost'],
+                        'cost' => (float) $receptor['cost'],
                     ];
                 },
                 (array) $data['receptors'],
@@ -64,7 +64,7 @@ final class SendBulkMessengerResponse
             message: (string) $data['message'],
             sendTime: isset($data['send_time']) ? new \DateTimeImmutable((string) $data['send_time']) : null,
             totalCount: (int) $data['total_count'],
-            totalCost: (int) $data['total_cost'],
+            totalCost: (float) $data['total_cost'],
             counts: $counts,
             profile: (string) $data['profile'],
             messenger: (string) $data['messenger'],
